# more diamond perf examples