--- a/performance-tests/tests/computedDiamondUpdates/computedDiamond1000UpdatesWith1000Dependencies.js
+++ b/performance-tests/tests/computedDiamondUpdates/computedDiamond1000UpdatesWith1000Dependencies.js
@@ -1,0 +1,3 @@
+var computedDiamondUpdate = require('../../scenarios/computedDiamondUpdate');
+
+module.exports = computedDiamondUpdate(1000, 1000);