--- a/performance-tests/tests/computedDiamondUpdates/computedDiamond1000UpdatesWith100Dependencies.js
+++ b/performance-tests/tests/computedDiamondUpdates/computedDiamond1000UpdatesWith100Dependencies.js
@@ -1,0 +1,3 @@
+var computedDiamondUpdate = require('../../scenarios/computedDiamondUpdate');
+
+module.exports = computedDiamondUpdate(1000, 100);